Export mapColumnToSchema function
I'm trying to create a mechanism to make it easy to write APIs for getting data collections (pagination, sort, filter). I use ElysiaJS and it is able to use Typebox types to validate request data, like query strings. I would like to have this function exported so it can be used to build other tools like the one I'm making. Won't always want to get the whole table. In my case I might want to specify fields that will be used as part of a output produced on a join. Please consider accepting this PR.