Handle type inference for postgres arrays
In the case of a pg array column such as tags: text('tags').array() when using the drizzle-zod plugin the type inference result of createSelectSchema() would be a zod schema with tags: z.ZodString rather than z.ZodArray(z.ZodString). At runtime everything seemed to work fine, but the typescript LSP would spit an error in the IDE. This should resolve that by allowing typescript to properly infer the nested array types.