Fix useLiveQuery potentially returning outdated data
This PR improves `useLiveQuery` to make it more robust.

Basically, if a component is re-rendered with a new query before the previous one finishes computing, and the second query is faster, the component can show outdated data (from the first query).

It makes it more robust by making sure that previous useEffect doesn't cause any state update.

As a side note, I think having `dips` as optional is a bit confusing and makes it unclear why changing the query doesn't change anything. It'd be great if there was a cache key for the query that could be used as a dependency of the useEffect instead.