Fix set transaction empty config error in pg-core
When an empty config object is supplied for postgres transactions, the client issues the query `set transaction ;` which causes a postgres parsing error for the pglite and postgres-js drivers. This commit uses the same approach seen in drizzle-orms mysql driver to only execute the `set transaction` sql when a non-empty config object is supplied to the transaction.