Add support for blob SQL type in relational queries
SQLite does not support selecting a blob type column in 'json_array()' and this function is used in relational queries.
Solve this by wrapping blob type columns in 'hex()' function to cast the blob to a hex string that we can parse back into a blob later.
Said 'later' being the 'mapFromDriverValue' step where inside this function we turn the hex string back into a buffer before passing it along.