Add AbortSignal to NodePgPreparedQuery
This signal is aborted from the NodePgSession object when an error event
occurs on the underlying node-postgres client object. By default, there
is no handler for such errors, meaning they will crash the Node.js
process.
The AbortSignal acts as a "buffer" for the error event, storing it and
turning it into a promise rejection that will then travel up the call
stack, allowing the code that invoked Drizzle to decide what to do with
it.