Allow migration without `CREATE SCHEMA` permission
Currently the migration uses `CREATE SCHEMA IF NOT EXISTS`, however that requires permission to create a schema even if there's already one. This change allows the migrations to be run by a user without the `CREATE SCHEMA` permission as long as the schema has already been created in advance.

This includes a test, that failed before, but works now after the small change.