support absolute paths for drizzle migrations folder
The drizzle migrations `out` property assumes that the migrations folder will be specified as a relative path. Specifying it as an absolute path will fail.

This allows us to specify an absolute path.

Internally it changes the representation of the resources in the migrations folder into absolute paths, and resolves the `out` path in the drizzle config to an absolute one.

As a side-effect, during the preparation of migrations, paths are now reported as absolute, instead of relative.