Ex field and auto invalidation fixes for Upstash Cache (#4502)
* fix: ex field in constructor

* fix: handle autoInvalidation: false correctly

previously, we still tried to put the autoInvalidation: false queries in the composite table hashes, which didn't work correctly because the tables list is empty in put in that case. Now, we put these queries under a differet hash. So they aren't associated with the tables. This means they aren't deleted when the table is deleted.

* Fix dprint

---------

Co-authored-by: AndriiSherman <andreysherman11@gmail.com>