Add support for transaction behavior in LibSQL
This commit adds support for handling specified
transaction behavior in the LibSQL database Session.
As the documentation of transaction mode in LibSQL states
(https://tursodatabase.github.io/libsql-client-ts/types/TransactionMode.html)
is not compatible with any other implementation of
SQLite, namely:
- "immediate" mode is named as "write",
- "exclusive" mode is not supported,
- "read" mode is introduced, which is a LibSQL
  extension.

To achieve full type-safety of the `config`
parameter passed to `transaction` function, this
parameter must be typed in generic way. For that
purpose, additional types were introduced and
existing modified:
- `AbstractSQLiteTransactionConfig<T>` (sqlite-core),
  which provides base type for `config` parameter,
  that has `behavior` field typed as `T`,
- `SQLiteTransactionConfig` now implements
  interface `AbstractSQLiteTransactionConfig`,
  typed as `T = 'deferred' | 'immediate' | 'exclusive'`,
- `LibSQLTransactionConfig` provides typings for
  LibSQL-specific transaction modes.

Types related to Transactions, Session and Database
were also modified, as I couldn't find a way to
provide a default type for the introduced parameter
`TTransactionConfig`, without making the type-checker
angry. It is important to note, that outside of
`"drizzle-orm/src/libsql/session.ts"` which is only
file that contains runtime changes, all changes in
other files are made to satisfy new type constraints.

Additionally, the `batch` function now also supports
an optional transaction config, to align functionality
with the LibSQL client.

This commit relates to feature request - closes #1331.