Fix bug in mysql-proxy for empty results on execute path
I am debugging an issue with the Drizzle MySQL proxy. The callback is expected to return a object with three fields:
- rows
- insertId
- affectedRows

However, the code handling the callback here destructures the rows field and then expects the insertID and affectedRows field on that.

This throws.

The doc seem out of date here as well:
https://orm.drizzle.team/docs/connect-drizzle-proxy

It does mention that on certain `method` a different response is expected. However, I can't validate this in the code and the mode `execute` here is never mentioned in the docs.